requirements.txt
asgiref==3.10.0
boto3==1.40.71
botocore==1.40.71
channels==4.3.1
channels_redis==4.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
jmespath==1.0.1
msgpack==1.1.2
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
